+++ src/main/kotlin/calculator/ArithmeticOperator.kt
@@ -1,4 +1,4 @@
-package step2
+package calculator
 
 enum class ArithmeticOperator(
     val symbol: String,
