+++ src/main/kotlin/calculator/StringCalculator.kt
@@ -1,4 +1,4 @@
-package step2
+package calculator
 
 class StringCalculator(
     input: String,
